+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/Merchant/Group/MerchantGroupProductModifyRequest.cs
@@ -34,7 +34,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("group_id")]
         [System.Text.Json.Serialization.JsonPropertyName("group_id")]
-        public int GroupId { get; set; }
+        public long GroupId { get; set; }
 
         /// <summary>
         /// 获取或设置商品列表。
