+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/Merchant/Shelf/MerchantShelfDeleteRequest.cs
@@ -13,6 +13,6 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("shelf_id")]
         [System.Text.Json.Serialization.JsonPropertyName("shelf_id")]
-        public int ShelfId { get; set; }
+        public long ShelfId { get; set; }
     }
 }
