+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/Product/LimitedDiscount/ProductLimitedDiscountUpdateStatusRequest.cs
@@ -10,7 +10,7 @@
         /// </summary>
         [Newtonsoft.Json.JsonProperty("task_id")]
         [System.Text.Json.Serialization.JsonPropertyName("task_id")]
-        public int TaskId { get; set; }
+        public long TaskId { get; set; }
 
         /// <summary>
         /// 获取或设置抢购任务状态。
