+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/WxaApi/LiveBroadcast/Goods/WxaApiBroadcastGoodsGetApprovedResponse.cs
@@ -14,7 +14,7 @@
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("goodsId")]
                 [System.Text.Json.Serialization.JsonPropertyName("goodsId")]
-                public int GoodsId { get; set; }
+                public long GoodsId { get; set; }
 
                 /// <summary>
                 /// 获取或设置商品名称。
